Use a better URL validator. RegexValidator adds flags
valid.url relied on urlparse very heavily, but urlparse did not
throw an exception on some invalid addresses. urllib3 does this
better, but is not available in EL6.

Use the URL validator from django instead, and validate with a
regex.

To support this, give RegexValidator easier-to-set flags as part
of the class constructor.

Change-Id: I13429d0cc600446c8fd4c1be5e5a8752d6d992f7
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1231029
Signed-off-by: Ryan Barry <rbarry@redhat.com>