Unpersist and repersist snmpd.conf so it can be moved
EL 7.1 (and probably EL6 at some point) now tries to move
/var/lib/net-snmp/snmpd.conf to snmpd.conf.0, probably as a
backup. This fails, since it's persisted (and bind mounted).

Unpersist it, then re-persist it when changes which require
restarting snmpd happen. Also, rework the logic of enable_snmpd
a little so we're not starting and stopping and restarting the
service or touching config files to change a password through
snmpusm

Change-Id: I0ff002717a8bf2744efa995f2e7e09a60b1eaede
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1232412
Signed-off-by: Ryan Barry <rbarry@redhat.com>