recipe: Copy the libvirtd upstart into place for vdsm
We can not whitelist the libvirtd.upstart file in /usr/share/doc because we
don't install doc files (--excludedocs is used in the %packages sections).
The approach is now to bundle the file and copy it into the place where
vdsm expects it at build time.

In addition we still need the whitelisting in the minimizer part, because
otherwise the minimizer would remove it again.

Change-Id: I2e3575c5146cc75b72e47ab9ef85882a256ea545
Bug-Url:https://bugzilla.redhat.com/show_bug.cgi?id=1235350
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>