confirmation_page: Do not compare strings and numbers
In commit ab550a39 we added the validation for a warning
message but the string "MB" is available in one of
variables making the comparation impossible (numbers vs string).
The patch 45c479bf added "MB" string in the data_size
model making ab550a39 commit not able to finish.
This patch avoid such exception:
ValueError: invalid literal for int() with base 10

Change-Id: I700a4aed2f7c11cb5081377d65bc622e0a5e1f32
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>