init: Restart networking early during boot
Previously we were restarting networking quite late during the boot. The
reason to do this was that the restarting would also recognize the bind
mounted ifcfg files - which weren't available during the first start of
networking.
The key point here is that the network got restartde very late during
the boot process and could interfere with other services coming up.

Now we still do the restart, but much earlier, which should lead to fewer
collisions with other services.

Change-Id: I040f69d6df0a24a31f8057121029d04c172646da
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1240921
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>