password: Always persist passwd and shadow
Befor eit could happen that one of the files ended up to be unpersisted,
by catching the exception and putting the persistence in the finaly branch,
we ensure that the files (and the set passwords) are now always persisted.

In the worst case it could have happened that a user can not login anymore,
because the password got unpersisted.

Change-Id: I30d579672c49bd0d07d0b8f73308b3d6238da5d3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1246117
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>