functions: Fix persistence check
Previously a check to test the persistence of a file was broken and
always returned False. This lead to a situation where a files was registered
to be persisted more than once.
This could then lead - with a previously fixed bug - to the situation
where a file was bind mounted several times into place.

Change-Id: Ia8ef8bfb396761fcfbab366858a14865b9a75943
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1246117
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>