functions: Make sure a stored file is bind mounted
Previously the ovirt_store_config function was not correctly bind
mounting files into the target, if the contents of that file were unchanged
compared to the laready persisted file. This could cause a situation
where changes done to a file after ovirt_store_config was called, were lost.
This is especially annoying if those were newly set passwords.

Change-Id: If26579837d4b91e96f785d79ec6ad07201735747
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1246117
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>