update: Make sure to use the right kernel modules
The initramfs rebuild suffered from the fact that in the Engine upgrade flow, the running
kernel and the modules available in /lib/modules did not match the updated kernel version,
this is because the new rootfs would only become effective on the next boot.
This is a problem because dracut has assumptions.

This patch does some tricks - mainly mounting /lib/modules and /boot into the right
places - to enable dracut to work nicely.

Change-Id: Ie893e722e24b1dd2e8e41970e16091d83c65dad8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1270228
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>