Persist /var/lib/rhsm out of the box
RHSM entitlements and Satellite6 activations with activation keys
expect to be able to write data in /var/lib/rhsm.

Since this occurs simultaneously with subscription-manager, it
cannot happen after subscription-manager runs. We could run it
beforehand, but there is no activationkey integration in the TUI,
which leaves little recourse other than using the shell.

This should be transparent to users. Persist it at install time.

Change-Id: I8f39f7f5d738fb5f9d372823c55b25d8f190f2fc
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1275597
Signed-off-by: Ryan Barry <rbarry@redhat.com>