utils: Suppress failed unmount errors
Previously failed unmounts were ignored (which is not good) and an error
was shown on the screen.
With this patch, the failure is still ignored, but the error is only
getting logged as a warning, and not as an error anymore.

The failure is still ignored, because it's not the right time to change
the behavior of th Mount().unmount() call.

Change-Id: I7b7d195b12203cec0944889ce9c5819d33d40eab
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1275962
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>