recipe: Create a vdsm-like multipath.conf
Previously we were creating a quite sparse multipath.conf compared
to the default one created by vdsm.
With this patch we ask vdsm-tool during build to create the multipath.conf
the assumption is that this will avoid bugs when the host is doing
the switchroot from initramfs to userspace.

One problem which could occur is that the accepted devices change
because the device filters are different between the two config files.

Change-Id: I450b33b2f5e2325b5850bc05c4a9691b45df4964
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1275956
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>