Be smarter about failed RHNSM/Satellite registrations
Check subscription-manager when the page is loaded so we can keep
the selection on the radio widgets correct (instead of clearing
rhntype from the model), but catch unusual exit codes to reflect
whether it's not registered at all or not subscribed.

Also, auto attaching works now, so try to auto-attach in the
model.

Since subscription-manager doesn't return meaningful exit codes
(just 1 for all possible errors I've seen), check the output from
the exception.

Change-Id: Ied4c2774ab4c36b1d612fbf765f87e315adf73ac
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1280215
Signed-off-by: Ryan Barry <rbarry@redhat.com>