mount_liveos: Check a few more places, don't break to much
This is a conservative patch to fix mount_liveos on el7.
It seems that a device can not be mounted more than once anymore.
But the previous mount_liveos logic assumed this.

Instead of mounting a device multiple times, bind mounts can be used.
The logic has been updated to use bind mounts, if well known places
are already mounted.
Otherwise the old logic is used.

The old logic was also updated to complete the loop over mapper
devices.

Change-Id: I600f70aec8ad8054ec93770a9c8dc410ac2bb9cb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1285607
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>