Call ConfigurureSubscriptionManager in rhn_model
All the logic for SAM and Satellite 6 got broken out into a new
class with the rewrite, but it was missed in the transaction. Add
it if the right model values are present.

Remove the ca cert when removing the old configs so we grab a new
one if the registration is changed from SAM to Satellite 6 or
something (unlikely, but happens in QE)

Change-Id: I14eb998cd1fb12d6b393232a17199aac55b6331b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1280217
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1276597
Signed-off-by: Ryan Barry <rbarry@redhat.com>