Remove hooks/on-shutdown/01-save-virtwho, add persistence to inst
Testing reveals that systemd's reverse ordering of services means
that shutdown hooks execute after /config is already umounted.

This appeared to work during development because the script was
tested (and triggered) before actually rebooting, so
/config/etc/virt-who.d/... already existed when persisting
individual (globbed) files.

Since /config is umounted, we need to just add this to the list
of files/folders which are persisted at install time.

Change-Id: I75df21a0e59fe8591a9989262344d1e872b278e2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1209356
Signed-off-by: Ryan Barry <rbarry@redhat.com>