installer: Use shared subtree for initramfs regeneration
It seems that someties (root cause still unclear) the bind mount
ended up to be a private subtree.
this would leave some of our config files mounted in the
bind omunted subtree, and would prevent removal in the host
tree.
Now we explicitly set rshared for the bind mounted subtrees.

Change-Id: I9c5cf2e95dad62565dd45371ffd0fc3351602888
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1263648
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>