Sometimes BridgedNIC gets called with a bridge. Be smarter
Previously, if BridgedNIC was called with an actual bridge as the
argument (ovirtmgmt, for example), it set null values for some
fields, such as the bridge nic, which caused the network to show
unconfigured when it actually was.

Add a method to Bridges() to find slaves.
Add BridgeNIC (essentially BridgedNIC but setting the ifname to
the bridge instead of the slave)

Called BridgedNIC only if it's a slave. If it's a bridge, call
BridgeNIC, and set the ifname appropriately.

Check /sys to find the actual slave nic

Change-Id: I8e99a015ff5904bfccd4c0d1655d5489776a28dd
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1280241
Signed-off-by: Ryan Barry <rbarry@redhat.com>
(cherry picked from commit 053b216d788964875c73bd8430a74f9742d17bfa)