Sometimes BridgedNIC gets called with a bridge. Be smarter
Previously, if BridgedNIC was called with an actual bridge as the
argument (ovirtmgmt, for example), it set null values for some
fields, such as the bridge nic, which caused the network to show
unconfigured when it actually was.

Add a method to Bridges() to find slaves.
Add BridgeNIC (essentially BridgedNIC but setting the ifname to
the bridge instead of the slave)

Called BridgedNIC only if it's a slave. If it's a bridge, call
BridgeNIC, and set the ifname appropriately.

Check /sys to find the actual slave nic

Change-Id: I9f7181d8fea099c6dce9f287c56d85b92afad068
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1280241
Signed-off-by: Ryan Barry <rbarry@redhat.com>
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>