Unset BOOTIF if it's already set (TUI upgrades)
If "upgrade" is appended, BOOTIF is already removed, but this does
not carry through to TUI upgrades, since "ugprade" is not set on
the cmdline.

Check whether it's already set (this will be unset in a base
image), and unset it if it is.

Change-Id: I7d64efb10d4f3bbf7d5a482c76675ed56c4eaf10
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1289028
Signed-off-by: Ryan Barry <rbarry@redhat.com>
(cherry picked from commit f32bbe934e16e80d7eaf4b6e9a1ab71998ac8957)