init: Fix the HOSTNAME check
HOSTNAME is checkd initially in the init script when the hostname arge
is read from the cmdline.
This breaks the HOSTNAME check further down, when it is read from the
sysconfig/network file.
The fix is to make the variale local before perfoming the check, if it's
local then it will only have the value from the sourced sysconfig/network
file, and not from further up.

Change-Id: Icac879a8c1f3e2269aa42c6e7c4268d82e4f03e3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1254580
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>