install.py: Handle multiple kernels for rebuild initramfs
Currently, if users try to add external kernel modules
using edit-node it will fail as the install.py cannot
handle such situation as it only expect one kernel in the image.

To avoid such issue, this patch adds a logic for such scenario
and detects the higher kernel version available.

Change-Id: I7787a9e3e6bc0b6c5ca7a257f8668a599357cd79
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1330632
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>